Refactor v2 parser tests
